+++ src/integrations/tg/tg_answers/message_keyboard_edit_answer.py
@@ -4,11 +4,17 @@ from integrations.tg.tg_answers.interface import TgAnswerInterface
 
 
 class TgKeyboardEditAnswer(TgAnswerInterface):
+    """Ответ с отредактированной клавиатурой."""
 
     def __init__(self, answer: TgAnswerInterface):
         self._origin = answer
 
     async def build(self, update) -> list[httpx.Request]:
+        """Пересобрать запросы к API к телеграмма.
+
+        :param update: Update
+        :return: list[httpx.Request]
+        """
         return [
             httpx.Request(
                 request.method,
